requirements.txt
Django>=5.0.1,<6.0
psycopg2>=2.9.9,<3.0.0
django-phonenumber-field[phonenumberslite]>=7.3.0,<8.0.0
